requirements.txt
absl-py==0.7.0
aniso8601==7.0.0
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
awscli==1.16.131
azure-common==1.1.22
azure-storage-blob==2.0.1
azure-storage-common==2.0.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.90
botocore==1.12.121
breadability==0.1.20
bs4==0.0.1
bz2file==0.98
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
cysignals==1.9.0
Cython==0.29.4
decorator==4.3.2
defusedxml==0.5.0
dill==0.2.9
docopt==0.6.2
docutils==0.14
en-core-web-sm==2.1.0
entrypoints==0.3
flake8==3.7.5
Flask==1.0.3
Flask-Cors==3.0.8
Flask-JWT-Extended==3.18.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
future==0.17.1
gast==0.2.2
gensim==3.7.1
googletrans==2.4.0
grpcio==1.19.0
h5py==2.9.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
imbalanced-learn==0.4.3
imblearn==0.0
imutils==0.5.2
include==0.2.2
ipykernel==5.1.0
ipython==7.2.0
ipython-autotime==0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
langdetect==1.0.7
lazy-object-proxy==1.3.1
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mementos==1.3.1
MiniSom==2.1.5
mistune==0.8.4
mock==2.0.0
moviepy==1.0.0
multiprocess==0.70.7
murmurhash==1.0.2
mysql-connector-python==8.0.15
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.3
nltk==3.4
notebook==5.7.4
numpy==1.16.1
oauthlib==3.0.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.3
passlib==1.7.1
pathos==0.2.3
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
pipdeptree==0.13.2
pipreqs==0.4.9
piptree==0.1.3
pkginfo==1.5.0.1
plac==0.9.6
pox==0.2.5
ppft==1.6.4.9
preshed==2.0.1
proglog==0.1.9
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
psutil==5.5.0
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycodestyle-magic==0.2.5
pycparser==2.19
pyenchant==2.0.0
pyfasttext==0.4.6
pyflakes==2.1.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
PyMySQL==0.9.3
pyparsing==2.3.1
PyPDF2==1.26.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.3
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
readme-renderer==24.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==3.4.2
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.20.1
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
setuptools-scm==1.15.7
shap==0.28.5
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.0
smote-variants==0.2.6
soupsieve==1.9.1
spacy==2.1.4
spark-parser==1.8.7
SQLAlchemy==1.3.5
srsly==0.0.7
stanfordcorenlp==3.7.0.2
statistics==1.0.3.5
sumy==0.7.0
tabulate==0.8.3
TBB==0.1
tensorboard==1.13.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textblob==0.15.2
Theano==1.0.4
thinc==7.0.4
torch==1.0.1.post2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
tweepy==3.7.0
twine==1.13.0
uncompyle6==3.3.0
urllib3==1.24.1
virtualenv==16.6.0
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.11.1
xdis==4.0.0
xlrd==1.2.0
yarg==0.1.9
# Editable install with no version control (yihan-constants==0.1)
-e /Users/yihanbao/Downloads/yihan_constants
# Editable install with no version control (yihan-helper==0.1)
-e /Users/yihanbao/Downloads/yihan_helper
